Move DAO interfaces into own folder
