+++ src/main/java/edu/byu/cs/controller/AuthController.java
@@ -2,7 +2,7 @@ package edu.byu.cs.controller;
 
 import edu.byu.cs.dataAccess.DaoService;
 import edu.byu.cs.dataAccess.DataAccessException;
-import edu.byu.cs.dataAccess.UserDao;
+import edu.byu.cs.dataAccess.daoInterface.UserDao;
 import edu.byu.cs.model.User;
 import edu.byu.cs.util.Serializer;
 import org.slf4j.Logger;
