+++ src/main/java/edu/byu/cs/dataAccess/daoInterface/SubmissionDao.java
@@ -1,5 +1,7 @@
-package edu.byu.cs.dataAccess;
+package edu.byu.cs.dataAccess.daoInterface;
 
+import edu.byu.cs.dataAccess.DataAccessException;
+import edu.byu.cs.dataAccess.ItemNotFoundException;
 import edu.byu.cs.model.Phase;
 import edu.byu.cs.model.Submission;
 import org.eclipse.jgit.annotations.NonNull;
