+++ src/main/java/edu/byu/cs/dataAccess/daoInterface/UserDao.java
@@ -1,5 +1,6 @@
-package edu.byu.cs.dataAccess;
+package edu.byu.cs.dataAccess.daoInterface;
 
+import edu.byu.cs.dataAccess.DataAccessException;
 import edu.byu.cs.model.User;
 
 import java.util.Collection;
