+++ src/main/java/edu/byu/cs/dataAccess/memory/SubmissionMemoryDao.java
@@ -2,7 +2,7 @@ package edu.byu.cs.dataAccess.memory;
 
 import edu.byu.cs.dataAccess.DataAccessException;
 import edu.byu.cs.dataAccess.ItemNotFoundException;
-import edu.byu.cs.dataAccess.SubmissionDao;
+import edu.byu.cs.dataAccess.daoInterface.SubmissionDao;
 import edu.byu.cs.model.Phase;
 import edu.byu.cs.model.Submission;
