+++ src/main/java/edu/byu/cs/util/SubmissionUtils.java
@@ -5,7 +5,7 @@ import edu.byu.cs.autograder.score.Scorer;
 import edu.byu.cs.dataAccess.DaoService;
 import edu.byu.cs.dataAccess.DataAccessException;
 import edu.byu.cs.dataAccess.ItemNotFoundException;
-import edu.byu.cs.dataAccess.SubmissionDao;
+import edu.byu.cs.dataAccess.daoInterface.SubmissionDao;
 import edu.byu.cs.model.Phase;
 import edu.byu.cs.model.Submission;
 import org.eclipse.jgit.annotations.NonNull;
